Hide Button Delete VatTax, IncomeTax, SalesTax

+++ src/modules/master/vat/view.js
@@ -8,12 +8,21 @@ export class View {
     constructor(router, service) {
         this.router = router;
         this.service = service;
+        this.canDelete = true;
+        this.isUsedInSalesTax = false;
     }
 
     async activate(params) {
         const decoded = Base64Helper.decode(params.id);
         var id = decoded;
         this.data = await this.service.getById(id);
+        const result = await this.service.getSalesTaxById(id);
+
+        this.isUsedInSalesTax = result === true;
+
+        if (this.isUsedInSalesTax) {
+            this.editCallback = null;
+        }
     }
 
     list() {
@@ -24,14 +33,37 @@ export class View {
       this.list();
     }
 
+    // editCallback(event) {
+    //     const encoded = Base64Helper.encode(this.data.Id);
+    //     this.router.navigateToRoute('edit', { id: encoded });
+    // }
+
+
     editCallback(event) {
+        if (this.isUsedInSalesTax) {
+            return;
+        }
+
         const encoded = Base64Helper.encode(this.data.Id);
         this.router.navigateToRoute('edit', { id: encoded });
     }
 
-    deleteCallback(event) {
+
+
+    // deleteCallback(event) {
+    //     this.service.delete(this.data)
+    //         .then(result => {
+    //             this.list();
+    //         });
+    // }
+
+    deleteCallback() {
+        if (this.isUsedInSalesTax) {
+            return;
+        }
+
         this.service.delete(this.data)
-            .then(result => {
+            .then(() => {
                 this.list();
             });
     }
